requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
beautifulsoup4=4.12.3=pypi_0
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.8.30=hf0a4a13_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.54.1=pypi_0
frozendict=2.4.6=pypi_0
fsspec=2024.9.0=pypi_0
html5lib=1.1=pypi_0
idna=3.10=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.7=pypi_0
libexpat=2.6.3=hf9b8971_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.46.1=hc14010f_0
libzlib=1.3.1=h8359307_2
lxml=5.3.0=pypi_0
markupsafe=3.0.1=pypi_0
matplotlib=3.9.2=pypi_0
mpmath=1.3.0=pypi_0
multitasking=0.0.11=pypi_0
ncurses=6.5=h7bae524_1
networkx=3.4.1=pypi_0
numpy=2.1.2=pypi_0
openssl=3.3.2=h8359307_0
packaging=24.1=pypi_0
pandas=2.2.3=pypi_0
peewee=3.17.6=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=pyh8b19718_1
platformdirs=4.3.6=pypi_0
pykan=0.2.6=pypi_0
pyparsing=3.2.0=pypi_0
python=3.12.7=h739c21a_0_cpython
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h92ec313_1
requests=2.32.3=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=pyhd8ed1ab_0
six=1.16.0=pypi_0
soupsieve=2.6=pypi_0
sympy=1.13.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.13=h5083fa2_1
torch=2.4.1=pypi_0
tqdm=4.66.5=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.44.0=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
yfinance=0.2.44=pypi_0
